Move service stacks to package, add fail indicator
